Adicionando foreign key para as relações, turma_id removido da tabela de disciplinas, escola_id removido da tabela de funcionarios, serie_id e disciplina_id removidos da tabela de turmas, gestor_id adicionado na tabela de escolas, id_escola removido da tabela de gestors

+++ database/migrations/2019_11_05_194637_create_alunos_table.php
@@ -17,7 +17,7 @@ class CreateAlunosTable extends Migration
             $table->bigIncrements('id');
             $table->date('nascimento');
             $table->string('filiacao', 200)->nullable();
-            $table->integer('pessoa_id');
+            $table->bigInteger('pessoa_id');
             $table->timestamps();
 
         });
