+++ database/migrations/2019_11_05_194954_create_gestors_table.php
@@ -16,8 +16,7 @@ class CreateGestorsTable extends Migration
         Schema::create('gestors', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('formacao');
-            $table->integer('funcionario_id');
-            // $table->integer('escola_id');
+            $table->bigInteger('funcionario_id');
             $table->timestamps();
         });
     }
